requirements.txt
fastapi
slack-sdk
python-dotenv
slack_bolt